+++ tests/topology/testSimpleExpander.cpp
@@ -209,7 +209,7 @@ using namespace DGtal::Z2i;
 bool testLayers()
 {
 
-  GradientColorMap<int> cmap_grad( 0, 30 );
+  GradientColorMap<size_t> cmap_grad( 0, 30 );
   cmap_grad.addColor( Color( 128, 128, 255 ) );
   cmap_grad.addColor( Color( 255, 255, 128 ) );
   cmap_grad.addColor( Color( 128, 255, 128 ) );
@@ -270,8 +270,7 @@ bool testLayers()
       ++it )
   std::cout << " " << *it;
       board << CustomStyle( expander.layer().className(), 
-          new CustomFillColor
-          ( cmap_grad( expander.distance() ) ) )
+          new CustomFillColor( cmap_grad( expander.distance() ) ) )
       << expander.layer();
       
       expander.nextLayer();
